+++ sublime_linter.py
@@ -15,8 +15,28 @@ from .lint.const import WARN_ERR, STATUS_KEY
 from .panel import panel
 
 
+def backup_old_settings():
+    """
+    If user settings file in old format exists is rename it and display message to user.
+
+    Thereby it is disabled and and backef up.
+    """
+    usr_dir_abs = os.path.join(sublime.packages_path(), "User")
+    settings_file = os.path.join(usr_dir_abs, "SublimeLinter.sublime-settings")
+    if os.path.exists(settings_file):
+        path = "Packages/User/SublimeLinter.sublime-settings"
+        settings = sublime.decode_value(sublime.load_resource(path))
+
+        if "user" in settings:
+            new_name = "SublimeLinter (old).sublime-settings"
+            new_path = os.path.join(usr_dir_abs, new_name)
+            os.rename(settings_file, new_path)
+            msg = "SublimeLinter\n\nYour settings have been backed up to:\n{}\nin Packages/User/".format(new_name)  # noqa: 501
+            sublime.message_dialog(msg)
+
+
 def plugin_loaded():
-    util.backup_old_settings()
+    backup_old_settings()
 
     persist.plugin_is_loaded = True
     persist.settings.load()
